Fixed test for empty folders on uninstall

+++ administrator/components/com_fabrik/com_fabrik.manifest.class.php
@@ -471,18 +471,18 @@ class Com_FabrikInstallerScript
 				case false:
 					/* The file itself will already be deleted */
 					/* Remove any empty folders in the tree */
-/*
+
 					$dir = $file;
 					foreach (array_reverse($pathParts) as $path) {
 						$dir = dirname($dir);	
 						if (Folder::exists($dir) === false) continue;
-						if (empty(Folder::files($dir))) {
+						if (empty(Folder::files($dir)) && empty(Folder::folders($dir))) {
 							if (Folder::delete($dir) === false) {
 								throw new RuntimeException("Failed to delete empty folder $dir.  Please check your permissions.");
 							}
 						}
 					}
-*/
+
 					break;
 			}
 		}
